pyproject.toml
[tool.poetry]
name = "adl-contract-admin"
version = "0.1.0"
description = "This package delivers functions to build the Analytics Dynasty League (ADL) Contract Admin Sheet and automatically calculates things like contract extensions."
authors = ["Christian Lohr <clohr9@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aniso8601 = "9.0.1"
anyio = "4.3.0"
apispec = "6.4.0"
apispec-webframeworks = "1.0.0"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
async-lru = "2.0.4"
attrs = "23.2.0"
automat = "22.10.0"
babel = "2.14.0"
beautifulsoup4 = "4.12.3"
bidict = "0.23.1"
binaryornot = "0.4.4"
bleach = "6.1.0"
blinker = "1.7.0"
boto3 = "1.34.34"
botocore = "1.34.67"
cachetools = "5.5.1"
certifi = "2024.2.2"
cffi = "1.16.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.1"
constantly = "23.10.4"
cookiecutter = "2.5.0"
debugpy = "1.8.1"
decorator = "5.1.1"
deepdiff = "6.7.1"
defusedxml = "0.7.1"
dnspython = "2.6.1"
et-xmlfile = "1.1.0"
executing = "2.0.1"
fastjsonschema = "2.19.1"
flask = "3.0.2"
flask-cors = "4.0.0"
flask-restful = "0.3.10"
flask-socketio = "5.3.6"
fqdn = "1.5.1"
gevent = "23.9.1"
gevent-websocket = "0.10.1"
gitignore-parser = "0.1.11"
google-auth = "2.38.0"
google-auth-oauthlib = "1.2.1"
greenlet = "3.0.3"
gspread = "6.1.4"
h11 = "0.14.0"
httpcore = "1.0.4"
httplib2 = "0.22.0"
httpx = "0.27.0"
hyperlink = "21.0.0"
idna = "3.6"
incremental = "22.10.0"
ipykernel = "6.29.3"
ipython = "8.22.2"
ipywidgets = "8.1.2"
isoduration = "20.11.0"
itsdangerous = "2.1.2"
jedi = "0.19.1"
jinja2 = "3.1.3"
jmespath = "1.0.1"
json5 = "0.9.22"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
jupyter = "1.0.0"
jupyter-console = "6.6.3"
jupyter-events = "0.9.0"
jupyter-lsp = "2.2.4"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
jupyter-server = "2.13.0"
jupyter-server-terminals = "0.5.2"
jupyterlab = "4.1.4"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.25.4"
jupyterlab-widgets = "3.0.10"
kthread = "0.2.3"
markdown = "3.5.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.20.2"
matplotlib-inline = "0.1.6"
mdurl = "0.1.2"
mistune = "3.0.2"
nbclient = "0.9.0"
nbconvert = "7.16.2"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
notebook = "7.1.1"
notebook-shim = "0.2.4"
numpy = "1.26.4"
oauth2client = "4.1.3"
oauthlib = "3.2.2"
openpyxl = "3.1.2"
ordered-set = "4.1.0"
overrides = "7.7.0"
packaging = "24.0"
pandocfilters = "1.5.1"
parso = "0.8.3"
passlib = "1.7.4"
platformdirs = "4.2.0"
polars = "0.20.15"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
pure-eval = "0.2.2"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pycparser = "2.21"
pygments = "2.17.2"
pymongo = "4.6.1"
pyparsing = "3.2.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-engineio = "4.9.0"
python-json-logger = "2.0.7"
python-slugify = "8.0.4"
python-socketio = "5.11.1"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
qtconsole = "5.5.1"
qtpy = "2.4.1"
referencing = "0.33.0"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rich = "13.7.1"
rpds-py = "0.18.0"
rpy2 = "3.5.15"
rsa = "4.9"
s3transfer = "0.10.1"
scipy = "1.15.1"
send2trash = "1.8.2"
setuptools = "69.2.0"
simple-websocket = "1.0.0"
six = "1.16.0"
sniffio = "1.3.1"
soupsieve = "2.5"
sqlalchemy = "2.0.25"
stack-data = "0.6.3"
taipy = "3.1.0"
taipy-config = "3.1.0"
taipy-core = "3.1.0"
taipy-gui = "3.1.0"
taipy-rest = "3.1.0"
taipy-templates = "3.1.0"
terminado = "0.18.0"
text-unidecode = "1.3"
tinycss2 = "1.2.1"
toml = "0.10.2"
tornado = "6.4"
traitlets = "5.14.1"
twisted = "23.10.0"
twisted-iocpsupport = "1.0.4"
types-python-dateutil = "2.8.19.20240311"
typing-extensions = "4.10.0"
tzdata = "2024.1"
uri-template = "1.3.0"
urllib3 = "2.2.1"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
werkzeug = "3.0.1"
widgetsnbextension = "4.0.10"
wsproto = "1.2.0"
zope-event = "5.0"
zope-interface = "6.2"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
